+++ handlers/users/__init__.py
@@ -1,1 +1,4 @@
-from . import handlers
+from . import assessments_handlers
+from . import send_files_handlers
+from . import menus_handlers
+from . import role_handlers
